# lib/api-zod/src/generated/types/favorite.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GamerVerse API
 * OpenAPI spec version: 0.1.0
 */

export interface Favorite {
  id: number;
  userId: string;
  gameSlug: string;
  gameName: string;
  /** @nullable */
  gameBackground?: string | null;
  gameRating: number;
  gameGenres: string;
  createdAt: Date;
}