# lib/api-zod/src/generated/types/genreSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GamerVerse API
 * OpenAPI spec version: 0.1.0
 */

export interface GenreSummary {
  id: number;
  name: string;
  slug: string;
}