# lib/api-zod/src/generated/types/platformEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GamerVerse API
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformEntryRequirements } from './platformEntryRequirements';
import type { PlatformSummary } from './platformSummary';

export interface PlatformEntry {
  platform: PlatformSummary;
  /** @nullable */
  requirements?: PlatformEntryRequirements;
}